requirements.txt
asgiref==3.10.0
diff-match-patch==20241021
Django==4.2.25
django-import-export==4.3.12
django-sequences==3.0
sqlparse==0.5.3
tablib==3.9.0
typing_extensions==4.15.0
python-dotenv==1.2.1
dj-database-url==3.0.1
whitenoise==6.11.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
wheel==0.38.1
